+++ test-server/test-ping.c
@@ -230,7 +230,7 @@ callback_lws_mirror(struct libwebsocket_context * this,
 			shift -= 8;
 		}
 
-		while (p - &pingbuf[LWS_SEND_BUFFER_PRE_PADDING] < size)
+		while (p - &pingbuf[LWS_SEND_BUFFER_PRE_PADDING] < (ptrdiff_t)size)
 			*p++ = 0;
 
 		gettimeofday(&tv, NULL);
@@ -267,13 +267,13 @@ callback_lws_mirror(struct libwebsocket_context * this,
 
 		if (n < 0)
 			return -1;
-		if (n < size) {
+		if (n < (int)size) {
 			lwsl_err("Partial write\n");
 			return -1;
 		}
 
 		if (flood &&
-			 (psd->ping_index - psd->rx_count) < (screen_width - 1))
+			 (psd->ping_index - psd->rx_count) < (unsigned long)(screen_width - 1))
 			fprintf(stderr, ".");
 		break;
